Integration of eForm with myDOCbase
Integration of eForm with myDOCbase:
>> The eForm will be converted to PDF and saved to the e-Filing system
once the workflow is completed with the eForm.
>> In Preamble & Postamble Emails and Conditional Emails, the eForm will
be converted to PDF and sent as attachment or an URL [for URL option,
the PDF file will be saved in a temporary physical folder].
>> The eForm can be edited by any User and it will be saved as a
different name if it is already used by any user.
>> The eForm's file size will be treated as 0 but when it will be
converted to PDF, the file size will be updated.